requirements.txt
faiss-cpu
sentence-transformers
streamlit==1.26.0
google-generativeai==0.3.0
python-dotenv==1.0.0
streamlit-lottie==0.0.5
pandas
requests